+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/onsd/wd
 
 go 1.16
+
+require (
+	github.com/kubernetes-sigs/kubebuilder v1.0.8 // indirect
+	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a // indirect
+)
